initial implementation of file deletions; needs tests

+++ write.go
@@ -5,7 +5,9 @@ import (
 	"fmt"
 	"go/format"
 	"io"
+	"io/ioutil"
 	"os"
+	"strings"
 )
 
 func writeType(w io.Writer, t *Type, opts options) {
@@ -58,7 +60,57 @@ func formatToBytes(b *bytes.Buffer) ([]byte, error) {
 	return formatted, nil
 }
 
+func getExistingGenFiles() (result map[string]bool) {
+	result = make(map[string]bool)
+
+	d, err := ioutil.ReadDir("./")
+
+	if err != nil {
+		panic(err)
+	}
+
+	for _, f := range d {
+		if strings.HasSuffix(f.Name(), "_gen.go") {
+			result[f.Name()] = true
+		}
+	}
+
+	return
+}
+
+func promptDeletions(packages []*Package, existing map[string]bool, input io.Reader, output io.Writer) (deletions []string, ok bool) {
+	for _, p := range packages {
+		for _, t := range p.Types {
+			delete(existing, t.FileName())
+		}
+	}
+
+	if len(existing) > 0 {
+		for f := range existing {
+			deletions = append(deletions, f)
+		}
+
+		fmt.Fprintf(output, "  This will delete previously-generated files: %s\n  Continue [y/n]? ", strings.Join(deletions, ", "))
+
+		var confirm string
+		fmt.Fscan(input, &confirm)
+		ok = confirm == "y"
+	} else {
+		ok = true
+	}
+
+	return
+}
+
 func writeFiles(packages []*Package, opts options) {
+	existing := getExistingGenFiles()
+	deletions, ok := promptDeletions(packages, existing, os.Stdin, os.Stdout)
+
+	if !ok {
+		fmt.Println("  operation cancelled")
+		return
+	}
+
 	for _, p := range packages {
 		for _, t := range p.Types {
 			file, err := os.Create(t.FileName())
@@ -80,5 +132,10 @@ func writeFiles(packages []*Package, opts options) {
 
 			fmt.Printf("  generated %s, yay!\n", t.Plural())
 		}
+	}
+
+	for _, f := range deletions {
+		os.Remove(f)
+		fmt.Printf("  deleted %s\n", f)
 	}
 }
